Implement all app features, code review fixes, and App Store preparation
Phase 1-B: Fix balance calculation bug (was using only 10 transactions),
create shared DataStore for cross-tab sync, fix Apple Sign-In nonce,
connect CurrencyManager to all views, fix Decimal precision, migrate
credentials to xcconfig, fix Supabase SDK version (3.0.0 → 2.43.0).

Phase 2: Budget management with period calculations and progress bars,
analytics with Swift Charts (BarMark cashflow, SectorMark categories),
real exchange rates via open.er-api.com with 1-hour cache.

Phase 3: Savings goals with progress rings and contributions,
subscription tracker with monthly totals, profile editing,
categories management, notification settings.

Phase 4: AI assistant with full-screen chat UI calling Supabase
edge functions, conversation history, follow-up suggestions.
Achievement system with badges, tiers, and category filters.

Phase 5: Premium paywall (StoreKit 2 stub for v1), 5-step onboarding
wizard, animated splash, savings snapshot on home tab.

Code review: Remove unnecessary @Observable from SupabaseManager,
extract Color(hex:) to dedicated extension, cache balance and category
lookups in DataStore via didSet, cache DateFormatters in ViewModels,
fix HomeSavingsSnapshotView missing .task modifier, add accessibility
labels to key UI elements.

App Store: Add Sign in with Apple entitlement, create PrivacyInfo.xcprivacy
manifest, add README.md with full project documentation.

+++ AkifiIOS/Extensions/Decimal+Currency.swift
@@ -10,8 +10,8 @@ extension Decimal {
 }
 
 extension Int64 {
-    /// Convert base amount (stored in minor units) to display value
-    var displayAmount: Double {
-        Double(self) / 100.0
+    /// Convert base amount (stored in minor units) to display Decimal
+    var displayAmount: Decimal {
+        Decimal(self) / 100
     }
 }
